# documentation/swagger.yaml
openapi: 3.1.0
info:
  title: Reto IBK API
  description: API de Productos y Pedidos
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: Productos
    description: API para gestión de productos
  - name: Pedidos
    description: API para gestión de pedidos con cálculo automático de descuentos
paths:
  /api/productos/{id}:
    get:
      tags:
        - Productos
      summary: Obtener producto por ID
      description: Retorna un producto específico por su ID
      operationId: getProductoById
      parameters:
        - name: id
          in: path
          description: ID del producto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Producto encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '404':
          description: Producto no encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '500':
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductoResponse'
    put:
      tags:
        - Productos
      summary: Actualizar producto
      description: Actualiza un producto existente
      operationId: updateProducto
      parameters:
        - name: id
          in: path
          description: ID del producto
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoRequest'
        required: true
      responses:
        '200':
          description: Producto actualizado exitosamente
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '404':
          description: Producto no encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '500':
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductoResponse'
    delete:
      tags:
        - Productos
      summary: Eliminar producto
      description: Elimina un producto del sistema
      operationId: deleteProducto
      parameters:
        - name: id
          in: path
          description: ID del producto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Producto eliminado exitosamente
        '404':
          description: Producto no encontrado
        '500':
          description: Error interno del servidor
  /api/pedidos/{id}:
    get:
      tags:
        - Pedidos
      summary: Obtener pedido por ID
      description: Retorna un pedido específico por su ID
      operationId: getPedidoById
      parameters:
        - name: id
          in: path
          description: ID del pedido
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pedido encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '404':
          description: Pedido no encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '500':
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PedidoResponse'
    put:
      tags:
        - Pedidos
      summary: Actualizar pedido
      description: Actualiza un pedido existente con nueva lista de productos. El total se recalcula automáticamente aplicando descuentos
      operationId: updatePedido
      parameters:
        - name: id
          in: path
          description: ID del pedido
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoRequest'
        required: true
      responses:
        '200':
          description: Pedido actualizado exitosamente con descuentos aplicados
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '400':
          description: Datos de entrada inválidos o productos no encontrados
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '404':
          description: Pedido no encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '500':
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PedidoResponse'
    delete:
      tags:
        - Pedidos
      summary: Eliminar pedido
      description: Elimina un pedido del sistema
      operationId: deletePedido
      parameters:
        - name: id
          in: path
          description: ID del pedido
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Pedido eliminado exitosamente
        '404':
          description: Pedido no encontrado
        '500':
          description: Error interno del servidor
  /api/productos:
    get:
      tags:
        - Productos
      summary: Obtener todos los productos
      description: Retorna una lista paginada de productos
      operationId: getAllProductos
      parameters:
        - name: page
          in: query
          description: Número de página (inicia en 0)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          description: Tamaño de página
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: Lista de productos obtenida exitosamente
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoResponse'
        '500':
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoResponse'
    post:
      tags:
        - Productos
      summary: Crear producto
      description: Crea un nuevo producto en el sistema
      operationId: createProducto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoRequest'
        required: true
      responses:
        '201':
          description: Producto creado exitosamente
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '500':
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductoResponse'
  /api/pedidos:
    get:
      tags:
        - Pedidos
      summary: Obtener todos los pedidos
      description: Retorna una lista paginada de pedidos
      operationId: getAllPedidos
      parameters:
        - name: page
          in: query
          description: Número de página (inicia en 0)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          description: Tamaño de página
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: Lista de pedidos obtenida exitosamente
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PedidoResponse'
        '500':
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PedidoResponse'
    post:
      tags:
        - Pedidos
      summary: Crear pedido
      description: 'Crea un nuevo pedido con lista de productos. El total se calcula automáticamente aplicando descuentos: 10% si total > $1000, 5% adicional si > 5 productos diferentes'
      operationId: createPedido
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoRequest'
        required: true
      responses:
        '201':
          description: Pedido creado exitosamente con descuentos aplicados
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '400':
          description: Datos de entrada inválidos o productos no encontrados
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '500':
          description: Error interno del servidor
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PedidoResponse'
components:
  schemas:
    ProductoRequest:
      type: object
      description: Datos del producto a crear
      properties:
        nombre:
          type: string
          minLength: 1
        precio:
          type: number
        stock:
          type: integer
          format: int32
      required:
        - nombre
        - precio
    ProductoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        precio:
          type: number
        stock:
          type: integer
          format: int32
    PedidoItemRequest:
      type: object
      properties:
        productoId:
          type: integer
          format: int64
        cantidad:
          type: integer
          format: int32
      required:
        - productoId
    PedidoRequest:
      type: object
      description: Datos del pedido con lista de productos y cantidades
      properties:
        fecha:
          type: string
          format: date
        estado:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/PedidoItemRequest'
          minItems: 1
      required:
        - items
    PedidoItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productoId:
          type: integer
          format: int64
        productoNombre:
          type: string
        cantidad:
          type: integer
          format: int32
        precioUnitario:
          type: number
        subtotal:
          type: number
    PedidoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fecha:
          type: string
          format: date
        total:
          type: number
        estado:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/PedidoItemResponse'