# src/main/java/mabmab/retoibk/reto/infrastructure/controller/dto/response/ProductoResponse.java
package mabmab.retoibk.reto.infrastructure.controller.dto.response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ProductoResponse {
    private Long id;
    private String nombre;
    private BigDecimal precio;
    private int stock;
}